Minecraft 1.7 For some reason the crypto stuff magically stopped working for me, so it is currently disabled.

+++ include/server/packet/packet_change_game_state.h
@@ -4,4 +4,4 @@ enum
 	CHANGE_GAME_MODE = 3
 };
 
-extern void packet_send_change_game_state(struct client *client, uint8_t reason, uint8_t gamemode);
+extern void packet_send_change_game_state(struct client *client, uint8_t reason, float gamemode);
